# backend/src/main/java/com/uniqdata/backend/project/ProjectUpdateDto.java
package com.uniqdata.backend.project;

import lombok.Data;

@Data
public class ProjectUpdateDto {
    private String title;
    private String description;
    private Project.ProjectStatus status;
    private Long escrowAmountXrp;
}