# swagger/backend-openapi.yaml
openapi: 3.0.3
info:
  title: Backend API (연구 서비스)
  description: |
    **서버 2 — 연구 서비스 전체.** 연구 CRUD, 참여자 관리, 대시보드.
    블록체인 처리가 필요할 때 내부에서 **Core(서버 1)** API를 호출합니다.
    - 참여 신청 → Core `POST /api/escrow` 호출 후 참여자 저장
    - 참여 철회 → Core `POST /api/escrow/cancel` 호출 후 비활성화
    - 대시보드 요약 → DB 집계 + Core `GET /api/summary` 결과 합침
  version: 1.0.0
  contact:
    name: UniQdata Backend

servers:
  - url: http://localhost:8080
    description: 로컬 개발
  - url: http://backend:8080
    description: Docker 내부

tags:
  - name: Projects
    description: 연구(프로젝트) CRUD. Wireframe 연구 목록/상세/생성/수정/삭제
  - name: Participants
    description: 참여 신청·철회 (내부에서 Core 에스크로 생성/취소 호출)
  - name: Dashboard
    description: 대시보드 KPI (DB + Core 에스크로 잔액)

paths:
  /api/v2/projects:
    get:
      tags:
        - Projects
      summary: 연구 목록 조회
      description: |
        Wireframe `GET /api/v2/projects` 대응.
        쿼리로 상태 필터, 페이징(limit, offset) 가능.
      operationId: listProjects
      parameters:
        - name: status
          in: query
          description: 상태 필터. all 또는 DRAFT, RECRUITING, COLLECTING, ANALYZING, COMPLETED
          schema:
            type: string
            enum: [all, DRAFT, RECRUITING, COLLECTING, ANALYZING, COMPLETED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
    post:
      tags:
        - Projects
      summary: 연구 생성
      description: 초안(DRAFT) 상태로 연구를 생성합니다.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '200':
          description: 생성된 연구
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  /api/v2/projects/{projectId}:
    get:
      tags:
        - Projects
      summary: 연구 상세 조회
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: 연구 없음
    patch:
      tags:
        - Projects
      summary: 연구 수정
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: 수정된 연구
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      tags:
        - Projects
      summary: 연구 삭제
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: 삭제 완료
        '404':
          description: 연구 없음

  /api/v2/projects/{projectId}/participants:
    get:
      tags:
        - Participants
      summary: 참여자 목록
      operationId: listParticipants
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Participant'

  /api/v2/projects/{projectId}/participants/by-address:
    get:
      tags:
        - Participants
      summary: 참여자 1명 조회 (주소 기준)
      operationId: getParticipantByAddress
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: participantAddress
          in: query
          required: true
          schema:
            type: string
          description: XRPL 참여자 지갑 주소
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '404':
          description: 참여자 없음

  /api/v2/projects/{projectId}/participants/enroll:
    post:
      tags:
        - Participants
      summary: 참여 신청
      description: |
        연구에 참여 신청. **내부에서 Core `POST /api/escrow` 호출** 후 참여자 저장.
        프로젝트의 escrowAmountXrp 만큼 에스크로 생성. participantAddress는 XRPL 지갑 주소.
      operationId: enrollParticipant
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantAddress]
              properties:
                participantAddress:
                  type: string
                  example: "rN7n7otQDd6FczFgLdlqtyMVrn3e1DjxvV"
      responses:
        '200':
          description: 참여 완료 (에스크로 생성됨)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '400':
          description: participantAddress 누락 또는 이미 참여 중
        '502':
          description: Core(블록체인 서버) 연동 실패
  /api/v2/projects/{projectId}/participants/withdraw:
    post:
      tags:
        - Participants
      summary: 참여 철회
      description: |
        **내부에서 Core `POST /api/escrow/cancel` 호출** 후 참여 비활성화.
        저장해 둔 escrowOwnerAddress, offerSequence로 취소 요청.
      operationId: withdrawParticipant
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantAddress]
              properties:
                participantAddress:
                  type: string
      responses:
        '200':
          description: 철회 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '400':
          description: 이미 철회됨 또는 참여자 없음
        '502':
          description: Core 에스크로 취소 실패

  /api/v2/dashboard/summary:
    get:
      tags:
        - Dashboard
      summary: 대시보드 KPI
      description: |
        Wireframe `GET /api/v2/dashboard/summary` 대응.
        total_projects, total_participants는 DB 집계, escrow_balance는 **Core GET /api/summary** 결과 사용.
      operationId: getDashboardSummary
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'

components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: 연구(프로젝트) ID

  schemas:
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, RECRUITING, COLLECTING, ANALYZING, COMPLETED]
        escrowAmountXrp:
          type: integer
          format: int64
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProjectCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        escrowAmountXrp:
          type: integer
          format: int64
          nullable: true

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, RECRUITING, COLLECTING, ANALYZING, COMPLETED]
        escrowAmountXrp:
          type: integer
          format: int64
          nullable: true

    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        total:
          type: integer
        kpi:
          type: object
          properties:
            total_projects:
              type: integer

    Participant:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        participantAddress:
          type: string
        escrowOwnerAddress:
          type: string
          description: Core createEscrow 응답. 취소 시 사용
        offerSequence:
          type: integer
          format: int64
          nullable: true
        escrowTxHash:
          type: string
          nullable: true
        active:
          type: boolean
        enrolledAt:
          type: string
          format: date-time
        withdrawnAt:
          type: string
          format: date-time
          nullable: true

    DashboardSummary:
      type: object
      description: DB 집계 + Core 에스크로 잔액
      properties:
        total_projects:
          type: integer
        total_participants:
          type: integer
        total_datapoints:
          type: integer
        escrow_balance:
          type: number
          description: Core /api/summary 의 escrow_balance_xrp
        escrow_wallet_address:
          type: string
        network:
          type: string